refactor: add getEmitSockets helper to get online target(subscriber) sockets
